hell freakin yeah! more bytes saved. i feel like a hero for bytes... saving their digital lives.
got rid of _qModule method altogether because it was only referenced in require anyway... so just put it as part of require.

abbreviated some of what was in that method (was unnecessary)

added a new property to the module object that stores whether tag has been added or not...
should prevent it from being reinserted inadvertently. plus there were some bugs i foresaw with dependencies causing this.

i think this could address alan's concerns today about deps not working as expected